+++ src/app/button/terra-button.component.ts
@@ -27,6 +27,7 @@ export class TerraButtonComponent
     @Input() inputTooltipPlacement:string; //top, bottom, left, right
     @Input() inputIsActive:boolean;
     @Input() inputIsFlagged:boolean;
+    @Input() inputIsDivider:boolean;
     @Output() outputClicked = new EventEmitter<any>();
     
     constructor()
@@ -35,6 +36,7 @@ export class TerraButtonComponent
         this.inputType = 'button';
         this.inputIsActive = false;
         this.inputIsFlagged = false;
+        this.inputIsDivider = false;
     }
     
     private click():void