+++ src/app/pager/data/terra-pager.parameter.interface.ts
@@ -1,4 +1,6 @@
-export interface TerraPagerParameterInterface
+import { TerraBaseParameterInterface } from '../../data/terra-base-parameter.interface';
+
+export interface TerraPagerParameterInterface extends TerraBaseParameterInterface
 {
     page?:number;
     itemsPerPage?:number;